In general, this update adds some basic thumbnail/photo support.
gaym.h: Added declaration of gaym_no_such_nick handler (701 from gay.com for a whois on someone 
not logged on)

msgs.c: Added thumbnail support, and images in info. Whenever a whois occurs, the thumbnail is 
also grabbed from gay.com (via a gaim_url_fetch) and then stuffed into a buddy_icon (at the fetch 
callback)

Also, if the info was triggered because of an explicit request for info, then the get-profile.txt 
info is used to pull the main picture from the gay.com server. Then, an info dialog is displayed 
with a picture and bioline. It could probably be cleaned up a lot.

There is the addition of a local struct, as well, for passing data amongst the callbacks.


parse.c: Just added msg_no_such_nick parsing. Probably didn't even need to, since it's just a 
different number on gay.com, but whatever.


gaym.c: 
*Added a member to gaym_conn to remember if the current whois request should trigger a 
userinfo dialog.
*Replaced NO_BUDDY_ICONS with buddy icon specifications in the prpl info struct
*Added a signal handler for conversation creation to trigger the retreival of a buddy icon. There 
is a corresponding handler function for this, as well (get_photo_info).