Re-organize chat room flags, and make a function for deciding what they should be:
This commit is experimental, subject to comment.  It currently uses the
following order:

   ICON        CODE MEANING
   ----------- ---- -----------------------------------------------
   founder.png   18 profile and g-rated photo 
   op.png        19 profile and x-rated photo
   halfop.png     8 profile only
   voice.png        unknown flag (we must need to do more analysis)
   [none]         0 no profile and no photo

While working on this, I wanted to more or less preserve the current order
but use halfop.png to mean profile but no photo (i.e. "half"), provide
a visual distinction between "profile without photo" and "nothing at all",
and provide a visual cue if there is some code we are not aware of.