Trying out a system that only uses the gaim buddy icon system for popup thumbnails. It's a hack, because of the way gaim manages buddy icons. In order for it to work for rooms, we need to create either a dummy conversation, or a dummy buddy list entry, otherwise, when you instantiate a buddy icon, it will be immediately destroyed.
This is because, the only API hook for creating a buddy icon calls create, and then calls update. 
If update does not find a conversation or blist node for the user, the icon is destroyed. You
can not call anything lower level than this function, because it's all static below that.

I'm going to experiment with transient conversations and buddy list entries.