NEW FEATURE: ROOM BROWSER! It's new and experimental. Don't be suprised if it does weird stuff. I'm still trying to make it perfect. Play around with it! Right click on a room in your buddy list and choose "Browse <roomname>"
The "Pref" column is the shortened version of user name that /WHO gives.
The "?" column shows an N if the WHO list and the NAMES list get out of sync.

There are probably some memory leaks, but hopefully nothing that opens hundreds of file pointers and leaves them dangling ;-)

One known bug that I realize as I write this: if you try to do a list on, say, #557=1 and #557=2 at the same time, bad things will happen. Probably not too bad, but not good, either. I need to update the hash and equals functions to fix this. 

File changes in gaym-extras are trivial, except for roombrowse.c, which was just a complete re-write.

cmds.c:
    Added the "WHO" command


msgs.c:
    Added a handler for the WHO command. It also manages a structure which is eventually sent back to the requester.


gaym.c:
    Add the gaym_get_room_namelist function for the plugin to call.
    Add a hash table for room lists.
    Add a function for hashing that allows #557=* to be the same as #557=1 or #557=8. Needs to be updated.  

weblogin.c:
    Got rid of some noisy debug messages.

gaym.h: 
    Add the GaymNamelist structure.
    Add the namelists hash to struct gaym_conn
    Add a prefix field to gaym buddies
    Add declaration for gaym_cmd_who
    Add declaration for get_room_namelist

helpers.c:
    Just commented out a debug msg I am not using anymore.